backend/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
PyPDF2